requirements.txt
asgiref==3.8.1
Django==5.2.2
django-rest-framework==0.1.0
djangorestframework==3.16.0
mysqlclient==2.2.7
pillow==11.2.1
psycopg==3.2.9
psycopg2==2.9.10
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
